update処理の実装(背景色変更を含む)、詳細検索の処理最適化

+++ update.cs
@@ -10,14 +10,28 @@ using System.IO;
 
 namespace 同人誌管理 {
     public partial class update : 同人誌管理.BookBase {
+
         public update() {
             InitializeComponent();
             this.Text = "更新";
+            this.genreComboBox.SelectedIndexChanged += new System.EventHandler(this.genreComboBox_SelectedIndexChanged);
+            this.originComboBox.SelectedIndexChanged += new System.EventHandler(this.originComboBox_SelectedIndexChanged);
+            this.mainChara.TextChanged += new System.EventHandler(this.mainChara_TextChanged);
+            this.authorsForm.TextChanged += new System.EventHandler(this.authorsForm_TextChanged);
+            this.circleForm.TextChanged += new System.EventHandler(this.circleForm_TextChanged);
+            this.titleForm.TextChanged += new System.EventHandler(this.titleForm_TextChanged);
+            this.bookShelf.SelectedIndexChanged += new System.EventHandler(this.bookShelf_SelectedIndexChanged);
+            this.storage.SelectedIndexChanged += new System.EventHandler(this.storage_SelectedIndexChanged_1);
+            this.yearForm.SelectedIndexChanged += new System.EventHandler(this.yearForm_SelectedIndexChanged);
+            this.monthForm.SelectedIndexChanged += new System.EventHandler(this.monthForm_SelectedIndexChanged);
+            this.dayForm.SelectedIndexChanged += new System.EventHandler(this.dayForm_SelectedIndexChanged);
+            this.all.CheckedChanged += new System.EventHandler(this.all_CheckedChanged);
+            this.r15.CheckedChanged += new System.EventHandler(this.r15_CheckedChanged);
         }
 
         public string[] resultArray = new string[0];
         public int currentIndex;
-        private string fileTitle;
+        private string beforeTitle;
 
         private void readUp_Click(object sender, EventArgs e) {
             if (currentIndex > 0) {
@@ -49,8 +63,7 @@ namespace 同人誌管理 {
             SQLiteConnect.Excute(query, ref reader);
             if (reader != null) {
                 reader.Read();
-                titleForm.Text = reader["タイトル"].ToString();
-                fileTitle = reader["タイトル"].ToString();
+                titleForm.Text = beforeTitle = reader["タイトル"].ToString();
                 circleForm.Text = reader["サークル"].ToString();
                 authorsForm.Text = reader["作者"].ToString();
                 originComboBox.Text = reader["作品"].ToString();
@@ -59,11 +72,11 @@ namespace 同人誌管理 {
                 yearForm.Text = date.Substring(0, 4);
                 monthForm.Text = date.Substring(4, 2);
                 dayForm.Text = date.Substring(6, 2);
-                switch (reader["年齢"].ToString()) {
-                    case "all": all.Checked = true; break;
-                    case "r15": r15.Checked = true; break;
-                    case "r18": r18.Checked = true; break;
+                foreach (RadioButton rb in ageLimit.Controls) {
+                    if (rb.Tag.ToString() == reader["年齢"].ToString())
+                        rb.Checked = true;
                 }
+
                 placeID = int.Parse(reader["place_ID"].ToString());
                 //ここでstorage_SelectedIndexChangedが動いてしまうので処理先で無理矢理動かさない
                 storage.Text = reader["場所"].ToString();
@@ -75,31 +88,23 @@ namespace 同人誌管理 {
                 //場所名変更時に変更出来なかった本棚一覧を改めて動かす
                 storage_SelectedIndexChanged(sender, e);
 
-                /*
-                query = "SELECT place_name FROM t_storage WHERE place_ID = " + placeID;
-                SQLiteConnect.Excute(query, ref reader);
-                storage.Text = reader["place_name"].ToString();
-                /*  場所の仕様変更の為一時コメントアウト
-                switch (reader[].ToString()) {
-                    case "house": house.Checked = true; break;
-                    case "hometown": hometown.Checked = true; break;
-                }
-                */
-
                 //イメージ表示
                 string imageFilePath = @"Thumbnail\" + idForm.Text + "_" + titleForm.Text + ".jpg";
                 //MessageBox.Show(imageFilePath);
                 if (File.Exists(imageFilePath)) {
                     pictureBox.Image = Image.FromFile(imageFilePath);
-                    imageFilePath = "処理済み";
                 }
-                
-                reader.Close();
-                SQLiteConnect.conn.Close();
+                else {
+                    pictureBox.Image = Image.FromFile(@"Thumbnail\NoImage.jpg");
+                }
+                restoreColors();
             }
         }
 
         private void dealing_Click(object sender, EventArgs e) {
+            //空白箇所の提言
+            if (checkNullForm()) return;
+
             string query;
 
             //UPDATE処理
@@ -115,15 +120,9 @@ namespace 同人誌管理 {
                     query += rb.Tag.ToString();
             }
             query += "',main_chara = '" + mainChara.Text + "'";
-            query += "," + "place = '";
-            /*  場所の仕様変更の為一時コメントアウト
-            if (house.Checked == true) {
-                query += "house";
-            } else if (hometown.Checked == true) {
-                query += "hometown";
-            }
-            */
-            query += "' WHERE " + resultArray[currentIndex] + " = ID";
+            query += "," + "place_ID = " + placeID;
+            query += "," + "bookShelf_ID = (SELECT bookShelf_ID FROM t_house_shelf WHERE shelf_name = '" + bookShelf.Text + "' )";
+            query += " WHERE " + resultArray[currentIndex] + " = ID";
 
             //t_doujinshi更新
             SQLiteConnect.Excute(query);
@@ -158,14 +157,81 @@ namespace 同人誌管理 {
             }
 
             //書き込み処理
-                string imageFileNameBefore = @"Thumbnail\" + idForm + "_" + fileTitle + ".jpg";
+                string imageFileNameBefore = @"Thumbnail\" + idForm + "_" + beforeTitle + ".jpg";
                 string imageFileNameAfter = @"Thumbnail\" + idForm + "_" + titleForm.Text + ".jpg";
             //ファイル名変更判定
-            if (titleForm.BackColor == Color.Cyan) {
+            if(File.Exists(imageFileNameBefore)){
+                if (titleForm.BackColor == Color.Cyan) {
                     File.Move(imageFileNameBefore, imageFileNameAfter);
+                }
             }
 
             MessageBox.Show("更新しました");
+            restoreColors();
+        }
+
+        //色をデフォルトに戻す
+        private void restoreColors() {
+            foreach (Control item in this.Controls) {
+                if (item is TextBox)
+                    item.BackColor = SystemColors.Control;
+                else if (item is Label)
+                    item.BackColor = SystemColors.Control;
+                else if (item is ComboBox)
+                    item.BackColor = SystemColors.Control;
+            }
+        }
+
+        private void titleForm_TextChanged(object sender, EventArgs e) {
+            titleForm.BackColor = Color.Cyan;
+        }
+
+        private void circleForm_TextChanged(object sender, EventArgs e) {
+            circleForm.BackColor = Color.Cyan;
+        }
+
+        private void authorsForm_TextChanged(object sender, EventArgs e) {
+            authorsForm.BackColor = Color.Cyan;
+        }
+
+        private void originComboBox_SelectedIndexChanged(object sender, EventArgs e) {
+            label4.BackColor = Color.Cyan;
+        }
+
+        private void genreComboBox_SelectedIndexChanged(object sender, EventArgs e) {
+            label5.BackColor = Color.Cyan;
+        }
+
+        private void yearForm_SelectedIndexChanged(object sender, EventArgs e) {
+            yearForm.BackColor = Color.Cyan;
+        }
+
+        private void monthForm_SelectedIndexChanged(object sender, EventArgs e) {
+            monthForm.BackColor = Color.Cyan;
+        }
+
+        private void dayForm_SelectedIndexChanged(object sender, EventArgs e) {
+            dayForm.BackColor = Color.Cyan;
+        }
+
+        private void storage_SelectedIndexChanged_1(object sender, EventArgs e) {
+            label7.BackColor = Color.Cyan;
+        }
+
+        private void bookShelf_SelectedIndexChanged(object sender, EventArgs e) {
+            label14.BackColor = Color.Cyan;
+        }
+
+        private void mainChara_TextChanged(object sender, EventArgs e) {
+            mainChara.BackColor = Color.Cyan;
+        }
+
+        private void all_CheckedChanged(object sender, EventArgs e) {
+            label6.BackColor = Color.Cyan;
+        }
+
+        private void r15_CheckedChanged(object sender, EventArgs e) {
+            label6.BackColor = Color.Cyan;
         }
     }
 }
